+++ src/framework/piral-base/src/loaders/v3/index.ts
@@ -35,7 +35,9 @@ export default function loader(entry: PiletV3Entry, options: DefaultLoaderConfig
 
     if (Array.isArray(app.styles)) {
       for (const style of app.styles) {
-        attachStyles(pilet, `${pilet.basePath}/${style}`);
+        const s = !style.startsWith('//') ? style.replace(/^\/+/, '') : style;
+        const url = new URL(s, pilet.basePath).toString();
+        attachStyles(pilet, url);
       }
     }