+++ src/plugins/piral-forms/src/index.ts
@@ -2,4 +2,5 @@ export * from './create';
 export * from './types';
 export * from './useForm';
 export * from './usePrompt';
+export * from './withFormHandler';
 export * from './withForm';