+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.10.0 (tbd)
 
 - Fixed usage of *piral.json* in monorepo scenarios for templating new pilets
+- Fixed deferred initialization for standalone Blazor applications using `piral-blazor`
 - Removed `always-auth` from all generated *.npmrc* files
 - Added support for Angular 21 in `piral-ng`