+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed initial layout decision in `piral-page-layouts` (#790)
 - Fixed resolution of style links in v3 loaders in `piral-base` (#802)
 - Removed `always-auth` from all generated *.npmrc* files
+- Added trusted publishing to CI/CD process
 - Added support for Angular 21 in `piral-ng`
 - Added `skipForm` config `createForm` pilet API function originating from `piral-forms` (#797)
 - Added `allowSubmitUnchanged` option to form creation in `piral-forms` (#798)